Restructure coverage tests to avoid Codecov patch false negatives
Move GCEventStore coverage scenarios into an existing test file and reduce GCEventStore token logic changes to directly exercised lines so patch coverage reflects executed code paths.

Made-with: Cursor
